Enabled cors on server and fetch from the caching server to client
