+++ client/.env
@@ -1,5 +1,1 @@
-REACT_APP_API_BASE_URL=https://bad-api-assignment.reaktor.com/
-
-REACT_APP_API_PRODUCTS_BY_CATEGORY_PATH=products/$category
-
-REACT_APP_API_AVAILABILITIES_BY_BRAND_PATH=availability/$brand
+REACT_APP_API_BASE_URL=http://localhost:3001/products
