--- client/src/utils/common/availabilitiesUtils.ts
@@ -1,7 +1,0 @@
-import { AvailabilitiesType, QuantityType } from "types/types";
-
-export const parseAvailabilityDataPayload = (payload: string): QuantityType => {
-    const result = payload.split("<INSTOCKVALUE>")[1].split("</INSTOCKVALUE>")[0] as QuantityType;
-
-    return result;
-};
