--- client/src/utils/common/productsUtils.ts
@@ -1,5 +1,0 @@
-import { ProductsApi } from "types/api";
-
-export const getProductsBrands = (products: ProductsApi) => {
-    return [...new Set(products.map((product) => product.manufacturer))];
-};
