+++ server/package.json
@@ -8,6 +8,7 @@
     },
     "license": "MIT",
     "devDependencies": {
+        "@types/cors": "^2.8.9",
         "@types/express": "^4.17.9",
         "@types/node-fetch": "^2.5.7",
         "ts-loader": "^8.0.14",
@@ -17,6 +18,7 @@
     },
     "dependencies": {
         "axios": "^0.21.1",
+        "cors": "^2.8.5",
         "dotenv": "^8.2.0",
         "express": "^4.17.1",
         "node-cache": "^5.1.2",
